refactor(layouts): split workspace layout into view and edit
