+++ src/layouts/layouts/SiteEditLayout/index.ts
@@ -1,2 +1,2 @@
-export * from "./SiteViewLayout"
 export * from "./SiteEditLayout"
+export * from "./SiteEditHeader"
